Replace all occurrences of long with int in the Heretic code.
Subversion-branch: /branches/raven-branch
Subversion-revision: 1270

+++ src/heretic/am_map.c
@@ -1118,7 +1118,7 @@ void DrawWuLine(int X0, int Y0, int X1, int Y1, byte * BaseColor,
         /* Y-major line; calculate 16-bit fixed-point fractional part of a
            pixel that X advances each time Y advances 1 pixel, truncating the
            result so that we won't overrun the endpoint along the X axis */
-        ErrorAdj = ((unsigned long) DeltaX << 16) / (unsigned long) DeltaY;
+        ErrorAdj = ((unsigned int) DeltaX << 16) / (unsigned int) DeltaY;
         /* Draw all pixels other than the first and last */
         while (--DeltaY)
         {
@@ -1147,7 +1147,7 @@ void DrawWuLine(int X0, int Y0, int X1, int Y1, byte * BaseColor,
     /* It's an X-major line; calculate 16-bit fixed-point fractional part of a
        pixel that Y advances each time X advances 1 pixel, truncating the
        result to avoid overrunning the endpoint along the X axis */
-    ErrorAdj = ((unsigned long) DeltaY << 16) / (unsigned long) DeltaX;
+    ErrorAdj = ((unsigned int) DeltaY << 16) / (unsigned int) DeltaX;
     /* Draw all pixels other than the first and last */
     while (--DeltaX)
     {
